+++ src/RecoveryServices/RecoveryServices.Backup.Helpers/Conversions/RetentionPolicyConversions.cs
@@ -627,6 +627,54 @@ namespace Microsoft.Azure.Commands.RecoveryServices.Backup.Helpers
 
         #endregion
 
+        #region Tiering Policy conversions 
+
+        /// <summary>
+        /// Helper function to convert service tiering policy from ps tiering policy.
+        /// </summary>
+        public static IDictionary<string, ServiceClientModel.TieringPolicy> GetServiceClientTieringPolicy(
+            TieringPolicy tierPolicy, bool isSmartTieringEnabled = false)
+        {
+            // do We need to call AzureWorkloadProviderHelper.ValidateTieringPolicy(tieringDetails) or TieringPolicy.Validate here ..   
+            IDictionary<string, ServiceClientModel.TieringPolicy> tieringPolicy = null;
+
+            // no validation is there in service if IaasVM duration is in Days 
+            // service also giving 0, Invalid - that doesn't sound meaningful 
+            // validate duration with Retention policy 
+            // validate duration type 
+            // validate duration 
+
+            if (isSmartTieringEnabled)
+            {
+                tieringPolicy = new Dictionary<string, ServiceClientModel.TieringPolicy>();
+                ServiceClientModel.TieringPolicy newTeringPolicy = new ServiceClientModel.TieringPolicy();
+                if (tierPolicy != null)
+                {
+                    tierPolicy.Validate();
+
+                    newTeringPolicy.TieringMode = (tierPolicy.TieringMode == TieringMode.TierAllEligible) ? ServiceClientModel.TieringMode.TierAfter.ToString() : tierPolicy.TieringMode.ToString();
+                    newTeringPolicy.DurationType = tierPolicy.TierAfterDurationType;
+                    newTeringPolicy.Duration = tierPolicy.TierAfterDuration;
+
+                    tieringPolicy.Add(ServiceClientModel.RecoveryPointTierType.ArchivedRP.ToString(), newTeringPolicy);
+                }
+                else // DO NOT TIER
+                {
+                    newTeringPolicy.TieringMode = ServiceClientModel.TieringMode.DoNotTier.ToString();
+                    tieringPolicy.Add(ServiceClientModel.RecoveryPointTierType.ArchivedRP.ToString(), newTeringPolicy);
+                }
+            }
+            else if (tierPolicy != null)
+            {
+                // can throw error that smart tiering is not Enabled for this subscription 
+            }
+                        
+            return tieringPolicy;
+        }
+
+        #endregion
+
+
         private static int GetIntegerFromNullableIntgerValue(int? value)
         {
             return (value.HasValue ? (int)value : default(int));